컴퓨터가 숫자를 임의선택하도록 함, 리팩토링작업
